+++ TowerDefense/src/byteDefense/model/ennemies/TrojanHorse.java
@@ -15,7 +15,7 @@ import byteDefense.utilities.BFS;
 
 public class TrojanHorse extends Ennemy {
 
-	private static final float ATTACK = 0;
+	private static final int ATTACK = 0;
 	private static final int DEFENSE = 0;
 	private static final int ATTACK_SPEED = 0;
 	private static final int ATTACK_RANGE = 0;
@@ -25,7 +25,7 @@ public class TrojanHorse extends Ennemy {
 		super(624, 432, bfsMap);
 	}
 
-	public float getAttack() {
+	public int getAttack() {
 		return ATTACK;
 	}
