+++ TowerDefense/src/byteDefense/model/towers/Antivirus.java
@@ -13,7 +13,7 @@ package byteDefense.model.towers;
 
 public class Antivirus extends Tower {
 
-	private static final float ATTACK = 0;
+	private static final int ATTACK = 0;
 	private static final int DEFENSE = 0;
 	private static final int ATTACK_SPEED = 0;
 	private static final int ATTACK_RANGE = 0;
@@ -23,7 +23,7 @@ public class Antivirus extends Tower {
 		super(x, y);
 	}
 
-	public float getAttack() {
+	public int getAttack() {
 		return ATTACK;
 	}
